refactor: replace old orchestrator with the new async pipeline
Drops the old per-step Scheduler/Buffer/Orchestrator loop and makes
the new async pipeline the only orchestrator. The new modules
(scheduler, engine, batcher, watcher, inference_admin, ckpt,
orchestrator) move into src/prime_rl/orchestrator/ and the dedicated
src/prime_rl/orch2 package is deleted.

Removed:
  src/prime_rl/orchestrator/{buffer,envs,eval_utils,event_loop_lag,
    inference_metrics,patches,utils,scheduler,orchestrator,ckpt}.py
  src/prime_rl/orch2/
  tests/unit/orchestrator/{test_buffer,test_scheduler,test_eval_scheduling}.py
  pyproject.toml: `orch2` entry point
  configs/rl.py: `use_orch2` flag
  entrypoints/rl.py: branch on use_orch2

Kept reusable utilities (advantage, filters, trajectories, vf_utils)
and env_server in place — they remained valid for the new orchestrator
and the trainer.

Config: kept all fields whose features we expect to add later (buffer,
collect_inference_metrics, heartbeat, optim, teacher_*, use_token_client,
seq_len, num_train_workers). They're accepted by the schema but ignored
by the orchestrator runtime; readers/validators that depend on them
(in rl.py and the trainer) are unchanged.

Smoke-tested: examples/reverse_text/rl.toml --max-steps 6 finishes
cleanly through the merged orchestrator.
