+++ src/prime_rl/entrypoints/rl.py
@@ -266,13 +266,12 @@ def rl_local(config: RLConfig):
             )
 
         # Start orchestrator process
-        orchestrator_bin = "orch2" if config.use_orch2 else "orchestrator"
         orchestrator_cmd = [
-            orchestrator_bin,
+            "orchestrator",
             "@",
             (config_dir / ORCHESTRATOR_TOML).as_posix(),
         ]
-        logger.info(f"Starting orchestrator process ({orchestrator_bin})")
+        logger.info("Starting orchestrator process")
         logger.debug(f"Orchestrator start command: {' '.join(orchestrator_cmd)}")
         with open(log_dir / "orchestrator.log", "w") as log_file:
             orchestrator_process = Popen(
