+++ src/prime_rl/orchestrator/batcher.py
@@ -13,9 +13,9 @@ from prime_rl.configs.orchestrator import (
     StepBatching,
     TokensBatching,
 )
-from prime_rl.orch2.ckpt import CkptManager, OrchState
-from prime_rl.orch2.engine import Group
 from prime_rl.orchestrator.advantage import AdvantageInputs, default_advantage_fn
+from prime_rl.orchestrator.ckpt import CkptManager, OrchState
+from prime_rl.orchestrator.engine import Group
 from prime_rl.orchestrator.filters import RolloutFilter, apply_filters
 from prime_rl.orchestrator.trajectories import interleave_rollout, pretokenize_rollout_trajectory
 from prime_rl.orchestrator.vf_utils import get_completion_len
