+++ src/prime_rl/orchestrator/ckpt.py
@@ -1,3 +1,4 @@
+import shutil
 import time
 from dataclasses import asdict, dataclass
 from pathlib import Path
@@ -5,92 +6,77 @@ from pathlib import Path
 import torch
 
 from prime_rl.configs.orchestrator import CheckpointConfig
-from prime_rl.orchestrator.buffer import Buffer
 from prime_rl.utils.logger import get_logger
-from prime_rl.utils.utils import get_ckpt_dir, get_step_path
+from prime_rl.utils.pathing import get_all_ckpt_steps, get_ckpt_dir, get_step_path
 
 
 @dataclass
-class Progress:
+class OrchState:
+    """Minimal orchestrator state needed to resume cleanly. Engine version +
+    watcher cursor are recoverable from disk on the next watcher tick, so we
+    don't persist them. Round-robin scheduler index resets to 0 on resume —
+    minor reordering, no correctness loss."""
+
     step: int = 0
-    total_tokens: int = 0
-    total_samples: int = 0
-    total_problems: int = 0
+    last_eval_step: int = 0
 
 
-class CheckpointManager:
-    """Utility class to save and load orchestrator checkpoints to resume orchestrator."""
+class CkptManager:
+    """Persist and restore the orch state under
+    `<output_dir>/checkpoints/step_N/orchestrator/state.pt` so trainer + orch
+    ckpts at the same step share a parent dir."""
 
     def __init__(self, output_dir: Path, config: CheckpointConfig):
         self.config = config
         self.ckpt_dir = get_ckpt_dir(output_dir)
         self.logger = get_logger()
 
-    def get_ckpt_path(self, step: int) -> Path:
+    def _path(self, step: int) -> Path:
         return get_step_path(self.ckpt_dir, step) / "orchestrator"
 
-    def save_to_path(
-        self,
-        ckpt_path: Path,
-        progress: Progress,
-        buffer: Buffer,
-    ):
-        self.logger.debug(f"Saving orchestrator checkpoint to {ckpt_path}")
-        start_time = time.perf_counter()
+    def save(self, state: OrchState, step: int) -> None:
+        path = self._path(step)
+        path.mkdir(parents=True, exist_ok=True)
+        t0 = time.perf_counter()
+        with open(path / "state.pt", "wb") as f:
+            torch.save({"state": asdict(state)}, f)
+        self.logger.info(f"Saved orch ckpt step {step} in {time.perf_counter() - t0:.2f}s")
+        self._prune()
 
-        # Save progress
-        with open(ckpt_path / "progress.pt", "wb") as f:
-            torch.save({"progress": progress}, f)
+    def load(self, step: int) -> OrchState:
+        path = self._path(step)
+        f = path / "state.pt"
+        if not f.exists():
+            raise FileNotFoundError(f"orch checkpoint missing at {f}")
+        with open(f, "rb") as fp:
+            data = torch.load(fp, weights_only=False)
+        return OrchState(**data["state"])
 
-        # Save buffer
-        buffer.save(ckpt_path / "buffer")
+    def latest_step(self) -> int | None:
+        if not self.ckpt_dir.exists():
+            return None
+        steps = [s for s in get_all_ckpt_steps(self.ckpt_dir) if (self._path(s) / "state.pt").exists()]
+        return steps[-1] if steps else None
 
-        self.logger.debug(f"Orchestrator checkpoint saved in {time.perf_counter() - start_time:.2f} seconds")
-
-    def load_from_path(self, ckpt_path: Path, progress: Progress, buffer: Buffer) -> None:
-        """Loads a checkpoint from a given path in-place."""
-        self.logger.debug(f"Loading checkpoint from {ckpt_path}")
-        start_time = time.perf_counter()
-
-        # Load progress
-        if self.config.skip_progress:
-            self.logger.info("Skipping progress loading from checkpoint")
-        else:
-            with open(ckpt_path / "progress.pt", "rb") as f:
-                state = torch.load(f, weights_only=False)
-
-            # Set progress in-place
-            for key, value in asdict(state["progress"]).items():
-                setattr(progress, key, value)
-
-        # Load buffer
-        if self.config.skip_buffer:
-            self.logger.info("Skipping buffer loading from checkpoint")
-        else:
-            buffer.load(ckpt_path / "buffer")
-
-        self.logger.debug(f"Orchestrator checkpoint loaded in {time.perf_counter() - start_time:.2f} seconds")
-
-    def load(self, progress: Progress, buffer: Buffer, step: int) -> None:
-        """Loads a checkpoint from a given path."""
-        ckpt_path = self.get_ckpt_path(step)
-        if not ckpt_path.exists():
-            raise FileNotFoundError(f"Checkpoint not found at {ckpt_path}")
-        self.load_from_path(ckpt_path, progress, buffer)
-
-    def save(
-        self,
-        progress: Progress,
-        buffer: Buffer,
-        step: int,
-    ) -> None:
-        """Saves the full checkpoint state for a specified step."""
-        ckpt_path = self.get_ckpt_path(step)
-        ckpt_path.mkdir(parents=True, exist_ok=True)
-        self.save_to_path(ckpt_path, progress, buffer)
+    def _prune(self) -> None:
+        # `keep_last` + `keep_interval` semantics match orch1: keep the last N
+        # ckpts always, plus any that are at a `keep_interval` multiple.
+        if self.config.keep_last is None and self.config.keep_interval is None:
+            return
+        steps = [s for s in get_all_ckpt_steps(self.ckpt_dir) if (self._path(s) / "state.pt").exists()]
+        keep: set[int] = set()
+        if self.config.keep_last is not None:
+            keep.update(steps[-self.config.keep_last :])
+        if self.config.keep_interval is not None:
+            keep.update(s for s in steps if s % self.config.keep_interval == 0)
+        for s in steps:
+            if s in keep:
+                continue
+            p = self._path(s)
+            if p.exists():
+                shutil.rmtree(p)
+                self.logger.debug(f"Pruned orch ckpt step {s}")
 
 
-def setup_ckpt_manager(output_dir: Path, config: CheckpointConfig | None) -> CheckpointManager | None:
-    if config is None:
-        return None
-    return CheckpointManager(output_dir, config)
+def setup_ckpt_manager(output_dir: Path, config: CheckpointConfig | None) -> CkptManager | None:
+    return CkptManager(output_dir, config) if config else None
