--- src/prime_rl/orchestrator/utils.py
@@ -1,169 +1,0 @@
-import asyncio
-import time
-from concurrent.futures import ThreadPoolExecutor
-from itertools import cycle
-from pathlib import Path
-from typing import Any
-
-import pandas as pd
-import verifiers as vf
-from rich.console import Console
-from rich.table import Table
-from verifiers.utils.client_utils import setup_openai_client
-
-from prime_rl.configs.orchestrator import OrchestratorConfig
-from prime_rl.transport import TrainingSample
-from prime_rl.utils.logger import get_logger
-from prime_rl.utils.utils import (
-    format_time,
-    get_broadcast_dir,
-    get_ckpt_dir,
-    get_step_path,
-)
-
-
-def set_default_executor(max_workers: int = 64) -> None:
-    """Scale the default asyncio thread pool so asyncio.to_thread has enough capacity."""
-    get_logger().info(f"Setting default executor to ThreadPoolExecutor(max_workers={max_workers})")
-    asyncio.get_event_loop().set_default_executor(ThreadPoolExecutor(max_workers=max_workers))
-
-
-def print_benchmark(history: dict[str, list[Any]]) -> None:
-    """
-    Print benchmark results as rich table. Shows formatted step time values.
-    First N rows show the per-step values, and the last row shows the mean,
-    std, min, and max values.
-    """
-    history.pop("step")
-    assert all(len(v) for v in history.values()), "All metrics must have logged the same number of steps"
-
-    # Turn metric history into pd.DataFrame
-    df = pd.DataFrame(dict(history.items()))
-    columns = {
-        "time/step": "Step Time",
-    }
-    df = df.rename(columns=columns)
-    df = df[list(columns.values())]
-    df = df.iloc[1:]  # Exclude first row
-
-    # Setup console
-    console = Console()
-    table = Table(title="Benchmark")
-
-    # Add columns
-    table.add_column("Step", justify="right")
-    for col in df.columns:
-        table.add_column(col, justify="center", style="magenta")
-
-    # Add formatted rows
-    formatted_df = pd.DataFrame(columns=df.columns)
-    formatted_df["Step Time"] = df["Step Time"].apply(format_time)
-    for step, row in formatted_df.iterrows():
-        table.add_row(*([str(step)] + [str(x) for x in row]))
-
-    # Separator
-    num_table_columns = 1 + len(df.columns)
-    table.add_row(*([""] * num_table_columns))
-
-    # Add row for formatted, aggregated statistics
-    mean_df = df.describe().loc[["mean", "std", "min", "max"], :]
-    formatted_mean_df = pd.DataFrame(columns=mean_df.columns)
-    formatted_mean_df["Step Time"] = mean_df["Step Time"].apply(format_time)
-    mean_row = ["Overall"] + formatted_mean_df.T.apply(
-        lambda row: f"{row['mean']} ± {row['std']} [{row['min']}, {row['max']}]", axis=1
-    ).tolist()
-    table.add_row(*mean_row)
-
-    # Display table
-    console.print(table)
-
-
-async def compute_teacher_logprobs(
-    clients: list[vf.ClientConfig],
-    model_name: str,
-    samples: list[TrainingSample],
-) -> list[list[float]]:
-    """Compute teacher model logprobs for a batch of training samples via prefill."""
-    from prime_rl.inference.vllm.serving_generate import GenerateResponse
-
-    async def _compute_single(client_config: vf.ClientConfig, sample: TrainingSample) -> list[float]:
-        client = setup_openai_client(client_config)
-
-        response = await client.post(
-            "/generate",
-            cast_to=GenerateResponse,
-            body={
-                "model": model_name,
-                "prompt_token_ids": sample.prompt_ids + sample.completion_ids,
-                "max_tokens": 1,
-                "temperature": 1.0,
-                "top_p": 1.0,
-                "prompt_logprobs": True,
-            },
-        )
-        return [0.0 if lp is None else float(lp) for lp in response.prompt_logprobs or []]
-
-    return await asyncio.gather(*[_compute_single(client, sample) for client, sample in zip(cycle(clients), samples)])
-
-
-def get_weight_dir(output_dir: Path, step: int, check_exists: bool = True, wait_timeout: int | None = None) -> Path:
-    """Get the weight directory for a given checkpoint step.
-
-    Args:
-        output_dir: The output directory for the run.
-        step: The checkpoint step.
-        check_exists: If True, raises FileNotFoundError if no weight directory exists.
-            If False, returns the broadcast directory path without checking existence
-            (useful for NCCL mode where weights are broadcasted, not stored on disk).
-        wait_timeout: Maximum time in seconds to wait for a stable directory to appear.
-            If None, no waiting is performed.
-    """
-    ckpt_weight_dir = get_step_path(get_ckpt_dir(output_dir), step) / "weight"
-    broadcast_weight_dir = get_step_path(get_broadcast_dir(output_dir), step)
-
-    def find_stable_dir() -> Path | None:
-        # For checkpoint weights, check STABLE file in parent directory (checkpoints/step_{step}/STABLE)
-        ckpt_step_dir = get_step_path(get_ckpt_dir(output_dir), step)
-        if (ckpt_step_dir / "STABLE").exists() and ckpt_weight_dir.exists():
-            return ckpt_weight_dir
-
-        # For broadcast weights, check STABLE file in the broadcast directory itself
-        if (broadcast_weight_dir / "STABLE").exists() and broadcast_weight_dir.exists():
-            return broadcast_weight_dir
-
-        return None
-
-    # Check immediately, then wait if needed
-    result = find_stable_dir()
-    if result is None and wait_timeout:
-        start_time = time.time()
-        while time.time() - start_time < wait_timeout:
-            time.sleep(1)
-            result = find_stable_dir()
-            if result:
-                break
-
-    if result:
-        return result
-    if not check_exists:
-        return broadcast_weight_dir
-
-    raise FileNotFoundError(f"No weight directory found for checkpoint step {step}")
-
-
-def setup_external_rollout_model(config: OrchestratorConfig, logger) -> tuple[Any, str, bool]:
-    """Resolve rollout client/model and whether policy updates should be enabled."""
-    rollout_client_config = config.client
-    rollout_model_name = config.model.name
-    enable_policy_updates = True
-
-    if config.teacher_rollout_model is not None:
-        rollout_client_config = config.teacher_rollout_model.client
-        rollout_model_name = config.teacher_rollout_model.model.name
-        enable_policy_updates = False
-        logger.info(
-            f"Using external teacher rollout model (base_url={', '.join(rollout_client_config.base_url)}, "
-            f"model={rollout_model_name})"
-        )
-
-    return rollout_client_config, rollout_model_name, enable_policy_updates
